refactor: gather kinetics related assets to kinetics/ folder test: after refactor and add hagrid x3d model all unit tests have passed
