+++ android/app/src/main/java/com/goliath/emojihub/repositories/local/X3dRepository.kt
@@ -22,11 +22,11 @@ class X3dRepositoryImpl @Inject constructor(
         const val DEFAULT_EMOJI_UNICODE = "U+0FE0F"
     }
     override fun createEmoji(videoUri: Uri): Pair<String, String>? {
-        val x3dModule = x3dDataSource.loadModule("efficient_x3d_xs_tutorial_float.pt")
+        val x3dModule = x3dDataSource.loadModule("kinetics/efficient_x3d_xs_tutorial_float.pt")
             ?: return null
         val (classNameFilePath, classUnicodeFilePath) = x3dDataSource.checkAnnotationFilesExist(
-            "hagrid_id_to_classname.json",
-            "hagrid_classname_to_unicode.json"
+            "kinetics/kinetics_id_to_classname.json",
+            "kinetics/kinetics_classname_to_unicode.json"
         )?: return null
         val videoTensor = loadVideoTensor(videoUri) ?: return null
         return predictEmojiClass(x3dModule, videoTensor, classNameFilePath, classUnicodeFilePath)
